# src/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Projeto Trainee Backend - API",
    "description": "Projeto para capacitação trainee de backend, com o objetivo de criar uma API para gerenciar projetos e tarefas."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearer_token": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "name": "bearer_token",
      "in": "header"
    }
  },
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/images/{path}/{filename}": {
      "get": {
        "tags": [
          "Images"
        ],
        "description": "Endpoint para retornar uma imagem",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Erro ao encontrar imagem."
          }
        }
      }
    },
...
  },
...
}